Remove unneeded comment. remove p variable.
p variable was just a placement because I was getting errors and I thought the error may have been a confusing with adding a string to a variable. So I had created another variable for just "p" to see if that was the error. Long story short, it wasn't the problem, so I reverted it to clean up the code.